chore: Test fixes
